create class MessageComponentWrapper
